requirements.txt
WTForms==2.1
numpy==1.19.4
pandas==1.1.5
imutils==0.5.4
Pillow==8.3.1
opencv-python==4.5.3.56
Werkzeug==2.0.1
opencv-python-headless==4.4.0.46
Flask-WTF==0.15.1
alembic==1.6.5
amqp==5.0.6
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
cachelib==0.2.0
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
greenlet==1.1.0
gspread==4.0.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
numpy==1.21.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
peewee==3.14.4
phonenumbers==8.12.28
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
supermemo2==2.0.0
twilio==6.62.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0